refactor: give the adaptive embedding path its own module
Adding the chunking and pooling work nearly doubled the provider module, which
is meant to be the trait, registry and shared types rather than a home for a
subsystem. The adaptive path moves to its own file unchanged, and the two
adapter modules that also crossed the size limit move their tests to sibling
files, matching the pattern anthropic already used.

No behaviour changes and no test lost: the same three thousand and ten tests
pass, and the architecture rules pass without adding a single exemption, since
an allowlist entry here would only have hidden the thing the rule exists to
report.
