+++ apps/desktop/src-tauri/src/commands/ai_provider/mod.rs
@@ -20,6 +20,7 @@ pub use crate::ipc_contracts::ai::{AiGenerateRequest, AiGenerateRequestMessage};
 
 mod anthropic;
 pub mod cli_agent; // pub: its registry/detection back the CLI-agent health probe
+mod embed; // adaptive chunk-and-mean-pool embedding machinery (R8 split — self-contained subsystem, own tests)
 mod gemini;
 pub mod ollama; // pub: its Ollama-only helpers back the local model list / health / embeddings
 mod ollama_cloud;
@@ -31,6 +32,7 @@ mod timeouts; // semantically-named per-request HTTP timeouts (pure extraction o
 
 use anthropic::AnthropicClient;
 use cli_agent::CliAgentClient;
+use embed::{embed_adaptive, ProviderEmbedAttempt};
 use gemini::GeminiClient;
 use ollama::OllamaClient;
 use ollama_cloud::OllamaCloudClient;
@@ -740,385 +742,6 @@ pub struct EmbeddingVector {
     pub space: EmbeddingSpace,
 }
 
-/// Char-boundary-safe truncation to at most `cap` chars (never splits a
-/// multi-byte char). Single pass: `char_indices().nth(cap)` finds the byte
-/// offset of the char *at* `cap` (i.e. the first char to drop). `Some` ⇒ the
-/// input exceeds `cap` chars, so slice there (a char-boundary offset); `None`
-/// ⇒ within cap, use as-is. Pure + unit-tested.
-fn truncate_chars(text: &str, cap: usize) -> &str {
-    match text.char_indices().nth(cap) {
-        Some((byte_offset, _)) => &text[..byte_offset],
-        None => text,
-    }
-}
-
-/// Never truncate an embedding input below this many chars — halving a
-/// context-length error down to nothing would still fail (an empty/near-empty
-/// embedding is meaningless) and would just multiply retries for no benefit.
-const EMBED_TRUNCATION_FLOOR_CHARS: usize = 500;
-
-/// Halve `current` for the next adaptive-truncation retry, clamped at
-/// [`EMBED_TRUNCATION_FLOOR_CHARS`]. `None` once `current` is already at or
-/// below the floor — the caller gives up rather than retry forever with the
-/// same (or a degenerate) length. Pure + unit-tested.
-fn next_truncation_len(current: usize) -> Option<usize> {
-    if current <= EMBED_TRUNCATION_FLOOR_CHARS {
-        None
-    } else {
-        Some((current / 2).max(EMBED_TRUNCATION_FLOOR_CHARS))
-    }
-}
-
-/// Whether an error MESSAGE (never a provider-specific error shape) reads as a
-/// context-length/input-too-long overflow — provider-agnostic on purpose, so a
-/// brand-new provider's own overflow wording is retried with zero code change
-/// here. Matches Ollama's `"...exceeds the context length"`, OpenAI's
-/// `"...maximum context length..."`/`"...too long..."`, Gemini's
-/// `"...exceeds the maximum number of tokens..."`, and the `friendly_api_error`
-/// 413 mapping (`"request too large"`). Pure + unit-tested.
-///
-/// Deliberately narrower than a bare `"too large"`/`"too long"` substring
-/// where that would over-match a same-wording RATE-LIMIT message (e.g. a
-/// "request too large for tokens-per-minute" 429 body) — `"request too
-/// large"`/`"payload too large"` are scoped to the actual over-length
-/// wording. This is NOT airtight (a provider could still phrase a rate-limit
-/// message with that exact phrase), but it is strictly narrower than the
-/// prior bare match and every known real provider wording still matches.
-///
-/// NOTE: deliberately does NOT attempt to fix this via a per-request
-/// `num_ctx`/`options.num_ctx` — Ollama ignores that per request; the model's
-/// context has to be baked into a derived model, which is out of scope here.
-fn is_context_length_error(msg: &str) -> bool {
-    let m = msg.to_ascii_lowercase();
-    m.contains("context length")
-        || m.contains("context_length_exceeded")
-        || m.contains("maximum context")
-        || m.contains("token limit")
-        || m.contains("too many tokens")
-        || m.contains("input length exceeds")
-        || m.contains("too long")
-        || m.contains("exceeds the maximum number of tokens")
-        || m.contains("request too large")
-        || m.contains("payload too large")
-}
-
-/// `AppHandle`-free shape of "embed this (already-truncated) text", so the
-/// adaptive-truncation retry loop (`embed_chunk_adaptive`) is unit-testable
-/// without a live `tauri::test` mock app — this crate has none (see the same
-/// note on `commands::ai::AnswerSearcher`). `embed_text` adapts the real
-/// `AiProvider::embed_with_usage` call (which needs `&AppHandle`) to this
-/// shape via `ProviderEmbedAttempt`.
-#[async_trait]
-trait EmbedAttempt {
-    async fn attempt(&self, text: &str) -> AppResult<(Vec<f64>, Usage)>;
-}
-
-struct ProviderEmbedAttempt<'a> {
-    app: &'a AppHandle,
-    client: &'a dyn AiProvider,
-    model: &'a str,
-}
-
-#[async_trait]
-impl EmbedAttempt for ProviderEmbedAttempt<'_> {
-    async fn attempt(&self, text: &str) -> AppResult<(Vec<f64>, Usage)> {
-        self.client
-            .embed_with_usage(self.app, self.model, text)
-            .await
-    }
-}
-
-/// Split `text` into consecutive, char-boundary-safe, non-overlapping chunks
-/// of at most `cap` chars each, preserving the WHOLE text (a naive single
-/// truncation silently drops everything past `cap` — indistinguishable from a
-/// complete embedding once stored, since the space tag is only
-/// `{provider, model, dim}`). Empty `text` yields exactly one empty chunk (so
-/// the caller still makes its usual single provider call rather than zero).
-/// Pure + unit-tested.
-fn split_into_chunks(text: &str, cap: usize) -> Vec<&str> {
-    if cap == 0 || text.is_empty() {
-        return vec![text];
-    }
-    let mut chunks = Vec::new();
-    let mut rest = text;
-    while !rest.is_empty() {
-        let piece = truncate_chars(rest, cap);
-        chunks.push(piece);
-        rest = &rest[piece.len()..];
-    }
-    chunks
-}
-
-/// Hard ceiling on the number of TOP-LEVEL chunks a single document is split
-/// into. On its own this does NOT bound total provider calls — a document
-/// needing several MAX_CHUNKS_PER_DOCUMENT-sized chunks, each independently
-/// re-discovering the provider's real per-request limit via
-/// `embed_chunk_adaptive`'s halving ladder, is still hundreds to thousands of
-/// calls. [`MAX_TOTAL_EMBED_ATTEMPTS`] is the real backstop on call volume;
-/// `embed_adaptive` also LEARNS the working cap once (across the whole
-/// document, not per chunk — see its own doc comment) so the halving ladder
-/// is normally paid only once, not once per chunk. There is no separate
-/// spend/budget cap in `crate::spend` for this — the exposure these two
-/// constants bound is call VOLUME/latency, not silent cost.
-const MAX_CHUNKS_PER_DOCUMENT: usize = 32;
-
-/// Hard ceiling on TOTAL provider calls (success + failure, across every
-/// chunk) for a single document's `embed_adaptive`. `MAX_CHUNKS_PER_DOCUMENT`
-/// alone bounds only the TOP-LEVEL chunk count, not the retries/sub-splits
-/// within them — measured up to ~4224 sequential calls for a 2MB document
-/// against a provider whose real limit was far below the nominal cap. This
-/// also bounds how long a single document can block its caller:
-/// `ai_reembed_all`'s cancellation check runs BETWEEN documents (in
-/// `commands::ai`), not within one, so an unbounded single-document embed
-/// makes Cancel unresponsive for however long that document takes. Once hit,
-/// `embed_adaptive` aborts with a clear error rather than continuing
-/// indefinitely — the caller's existing per-document failure handling
-/// (`ai_reembed_all` counts it as `failed` and moves on) takes it from there.
-const MAX_TOTAL_EMBED_ATTEMPTS: usize = 200;
-
-/// The per-chunk char cap `split_into_chunks` should actually use for `text`:
-/// `initial_cap`, UNLESS the document would need more than
-/// [`MAX_CHUNKS_PER_DOCUMENT`] chunks at that size — in which case the chunk
-/// size is grown just enough to stay within the cap. This NEVER drops text
-/// (unlike shrinking the document itself would): a larger chunk that still
-/// overflows the provider's real token window is caught and fully covered by
-/// `embed_chunk_adaptive`'s per-chunk halving-with-remainder, same as any
-/// other chunk — it just does a little more of that work. Pure + unit-tested.
-fn bounded_split_cap(total_chars: usize, initial_cap: usize) -> usize {
-    let cap = initial_cap.max(1);
-    let needed = total_chars.div_ceil(cap);
-    if needed <= MAX_CHUNKS_PER_DOCUMENT {
-        return cap;
-    }
-    let grown = total_chars.div_ceil(MAX_CHUNKS_PER_DOCUMENT);
-    tracing::warn!(
-        "embedding a {total_chars}-char document would need {needed} chunks at {cap} chars \
-         each — capping to {MAX_CHUNKS_PER_DOCUMENT} chunks of ~{grown} chars instead \
-         (adaptive per-chunk halving still covers the whole document, no text is dropped)"
-    );
-    grown
-}
-
-/// L2-normalize `v` in place (a no-op on an already-zero vector). Cosine
-/// similarity is normalization-invariant, so this doesn't change match
-/// quality — it keeps a mean-pooled vector on the same footing as a
-/// provider's own (typically already-normalized) single-call embedding.
-/// NOTE: this makes even a single-chunk embedding's stored VALUES
-/// byte-different from what the provider returned (same direction, unit
-/// length) — harmless for every current consumer, which all go through the
-/// scale-invariant `vector::cosine`, but a future DOT-PRODUCT (not cosine)
-/// consumer would silently mix normalized and un-normalized magnitudes if it
-/// read a pre-this-change stored row. Pure + unit-tested.
-fn l2_normalize(v: &mut [f64]) {
-    let norm = v.iter().map(|x| x * x).sum::<f64>().sqrt();
-    if norm > 0.0 {
-        for x in v.iter_mut() {
-            *x /= norm;
-        }
-    }
-}
-
-/// Embed ONE chunk (already at-or-under `initial_cap` chars) in full,
-/// adaptively re-sizing when the provider reports a context-length overflow
-/// — the per-chunk fallback for a token-dense language where even a
-/// `cap`-sized chunk overflows the model's real token window (a fixed char
-/// cap is fine for English but not for German, which tokenizes far denser).
-///
-/// Critically, a successful length that's SMALLER than the chunk is NOT the
-/// end — the unsent remainder is never discarded: once a length succeeds,
-/// that same length becomes the working cap for the rest of THIS chunk too
-/// (no repeating the failed larger caps), so the chunk's full text is always
-/// embedded, possibly as several vectors. `initial_cap` is clamped to the
-/// remaining length up front so a short remainder is never "halved" through
-/// cap values it was never actually sent at, and both the retry log and the
-/// give-up error name the length ACTUALLY sent, not the nominal cap. Gives up
-/// with the LAST provider error once truncating further would fall at/below
-/// [`EMBED_TRUNCATION_FLOOR_CHARS`], or once `attempts_used` (shared across
-/// every chunk of the same document — see `embed_adaptive`) would exceed
-/// [`MAX_TOTAL_EMBED_ATTEMPTS`]. Any non-context-length error returns
-/// immediately (never retried), abandoning whatever of the chunk is still
-/// unsent.
-///
-/// Returns the FINAL working cap alongside the vectors — the caller carries
-/// it forward as the STARTING cap for the next chunk, so the halving ladder
-/// is learned once per document, not re-paid by every chunk (see
-/// `embed_adaptive`).
-///
-/// `learned` (the returned/carried-forward value) is DELIBERATELY a
-/// SEPARATE variable from the per-attempt `cap` used to size each actual
-/// send. Only a genuine provider context-length REJECTION lowers `learned`;
-/// clamping `cap` down to whatever's left in a short final remainder (e.g.
-/// the chunk's last 200 chars) must NEVER also shrink `learned` — that was
-/// the bug: on a doc long enough to hit `bounded_split_cap`'s growth path,
-/// the tiny leftover tail of chunk N would poison the cap chunk N+1 starts
-/// at, permanently collapsing the "learned once" optimization into a
-/// near-worst-case per-piece ladder for the REST of the document (measured:
-/// a 300k-char document degraded to 1-char sends and hit
-/// `MAX_TOTAL_EMBED_ATTEMPTS` with barely 3% actually embedded).
-///
-/// `usage` is an OUTPUT parameter, not part of the return value, because it
-/// must accumulate a REAL provider-reported token count for every call that
-/// actually succeeded even when a LATER attempt in this same chunk fails —
-/// an early `return Err(..)` (the attempt ceiling, a non-context-length
-/// error, or the length-floor give-up) must never silently discard the
-/// spend already billed by the provider for the pieces that DID succeed.
-async fn embed_chunk_adaptive<A: EmbedAttempt>(
-    attempt: &A,
-    chunk: &str,
-    initial_cap: usize,
-    attempts_used: &mut usize,
-    usage: &mut Usage,
-) -> AppResult<(Vec<Vec<f64>>, usize)> {
-    let mut vectors = Vec::new();
-    let mut rest = chunk;
-    let mut learned = initial_cap.min(chunk.chars().count());
-    // `while first_pass || !rest.is_empty()`: an empty chunk (the whole
-    // document was empty) still makes exactly ONE provider call, matching
-    // every other caller's single-attempt behavior for empty input — it does
-    // NOT silently make zero calls just because `rest` starts empty.
-    let mut first_pass = true;
-    while first_pass || !rest.is_empty() {
-        first_pass = false;
-        // The length to actually send THIS pass — bounded by what's left in
-        // `rest` and by the best-known working size (`learned`). This local
-        // clamp must never feed back into `learned` itself (see doc comment).
-        let mut cap = learned.min(rest.chars().count());
-        loop {
-            if *attempts_used >= MAX_TOTAL_EMBED_ATTEMPTS {
-                return Err(AppError::Provider(format!(
-                    "embedding this document needed more than {MAX_TOTAL_EMBED_ATTEMPTS} \
-                     provider calls (context-length retries + chunk splits) — aborting rather \
-                     than continuing indefinitely"
-                )));
-            }
-            let truncated = truncate_chars(rest, cap);
-            let sent_len = truncated.chars().count();
-            *attempts_used += 1;
-            match attempt.attempt(truncated).await {
-                Ok((values, piece_usage)) => {
-                    usage.input_tokens =
-                        usage.input_tokens.saturating_add(piece_usage.input_tokens);
-                    usage.output_tokens = usage
-                        .output_tokens
-                        .saturating_add(piece_usage.output_tokens);
-                    vectors.push(values);
-                    rest = &rest[truncated.len()..];
-                    break;
-                }
-                Err(e) if is_context_length_error(&e.to_string()) => {
-                    match next_truncation_len(cap) {
-                        Some(smaller) => {
-                            tracing::warn!(
-                                "embed context-length error at {sent_len} chars — retrying at {smaller} chars"
-                            );
-                            cap = smaller;
-                            // A genuine discovery — this is the one place
-                            // `learned` may legitimately shrink.
-                            learned = smaller;
-                        }
-                        None => {
-                            return Err(AppError::Provider(format!(
-                                "embedding input is too long for this model even after truncating to \
-                                 {sent_len} characters: {e}"
-                            )));
-                        }
-                    }
-                }
-                Err(e) => return Err(e),
-            }
-        }
-    }
-    Ok((vectors, learned))
-}
-
-/// Embed the FULL `text` — chunk it at `initial_cap` chars
-/// ([`split_into_chunks`], bounded to at most [`MAX_CHUNKS_PER_DOCUMENT`] top-
-/// level chunks by [`bounded_split_cap`]), embed every chunk in full
-/// ([`embed_chunk_adaptive`] never drops a chunk's unsent remainder even
-/// when it has to sub-split on a context-length error), then mean-pool +
-/// L2-normalize every resulting vector into one ([`l2_normalize`]). This is
-/// what makes re-indexing self-tuning per document/language AND lossless: a
-/// naive single truncation silently drops everything past the cap while
-/// still tagging the result as a complete embedding (the space is only
-/// `{provider, model, dim}` — nothing marks a truncated vector as partial),
-/// so a long résumé would be indexed on roughly its first quarter with no
-/// way to tell. Usage is the REAL sum across every provider call actually
-/// made (chunking/sub-splitting a document legitimately costs proportionally
-/// more tokens than embedding it whole).
-///
-/// The working cap is LEARNED ONCE across the whole document, not
-/// re-discovered per chunk: `embed_chunk_adaptive` returns the cap it ended
-/// up succeeding at, and that becomes the STARTING cap for the next
-/// top-level chunk. Without this, a document needing `MAX_CHUNKS_PER_DOCUMENT`
-/// grown chunks would pay the full halving-ladder discovery cost on EVERY
-/// chunk — measured ~224 wholly wasted failed round-trips for a 2MB document
-/// against a provider whose real limit was far below the nominal cap; with
-/// this, only the first chunk pays that discovery cost. `attempts` is the
-/// running TOTAL across every chunk, bounded by [`MAX_TOTAL_EMBED_ATTEMPTS`].
-///
-/// `usage` is an OUTPUT parameter (see `embed_chunk_adaptive`'s doc comment)
-/// so a failure partway through a multi-chunk document — the attempt
-/// ceiling, or ANY error on chunk 5 of 8 — still leaves the REAL usage
-/// billed for chunks 1-4 available to the caller (`embed_text`) to record.
-/// Before this, `embed_text`'s `.await?` on the old `AppResult<(Vec<f64>,
-/// Usage)>` return discarded the whole `Usage` on any error, silently
-/// dropping already-billed spend from the ledger — a cost-VISIBILITY
-/// regression versus the pre-chunking code (which made exactly one call, so
-/// a failure billed nothing to begin with).
-async fn embed_adaptive<A: EmbedAttempt>(
-    attempt: &A,
-    text: &str,
-    initial_cap: usize,
-    usage: &mut Usage,
-) -> AppResult<Vec<f64>> {
-    let split_cap = bounded_split_cap(text.chars().count(), initial_cap);
-    let chunks = split_into_chunks(text, split_cap);
-    let mut pooled: Vec<f64> = Vec::new();
-    let mut expected_dim: Option<usize> = None;
-    let mut piece_count: usize = 0;
-    let mut attempts: usize = 0;
-    let mut learned_cap = split_cap;
-    for chunk in chunks {
-        let (pieces, cap_used) =
-            embed_chunk_adaptive(attempt, chunk, learned_cap, &mut attempts, usage).await?;
-        learned_cap = cap_used;
-        for values in pieces {
-            piece_count += 1;
-            match expected_dim {
-                None => {
-                    expected_dim = Some(values.len());
-                    pooled = values;
-                }
-                Some(d) if d == values.len() => {
-                    for (p, v) in pooled.iter_mut().zip(values.iter()) {
-                        *p += v;
-                    }
-                }
-                Some(_) => {
-                    // Same model/provider must always yield the same
-                    // dimension — a mismatch means the provider itself is
-                    // inconsistent, not something averaging can paper over.
-                    return Err(AppError::Provider(
-                        "embedding dimension changed between chunks of the same document"
-                            .to_string(),
-                    ));
-                }
-            }
-        }
-    }
-    // Mean-pool (a single piece is already its own mean — skip the no-op
-    // divide-by-one), then L2-normalize so a multi-piece pooled vector sits
-    // on the same footing as a provider's own single-call embedding.
-    if piece_count > 1 {
-        let n = piece_count as f64;
-        for p in pooled.iter_mut() {
-            *p /= n;
-        }
-    }
-    l2_normalize(&mut pooled);
-    Ok(pooled)
-}
-
 /// Embed `text` with an explicit provider/model, returning a space-tagged vector.
 /// Routes through the same `resolve` + capability + auth flow as chat, so there
 /// are no Ollama assumptions and no silent fallback.
@@ -1534,529 +1157,5 @@ mod tests {
     #[test]
     fn validate_model_cloud_requires_a_model() {
         assert!(ProviderId::OpenAi.validate_model("").is_err());
-    }
-
-    // ── Adaptive embedding truncation ───────────────────────────────────────
-
-    #[test]
-    fn truncate_chars_respects_char_boundaries() {
-        let text = "héllo wörld"; // multi-byte chars
-        assert_eq!(truncate_chars(text, 100), text);
-        assert_eq!(truncate_chars(text, 1), "h");
-        assert_eq!(truncate_chars(text, 2), "hé");
-    }
-
-    #[test]
-    fn next_truncation_len_halves_down_to_the_floor_then_gives_up() {
-        assert_eq!(next_truncation_len(8000), Some(4000));
-        assert_eq!(next_truncation_len(4000), Some(2000));
-        assert_eq!(next_truncation_len(2000), Some(1000));
-        assert_eq!(next_truncation_len(1000), Some(500));
-        assert_eq!(next_truncation_len(500), None);
-        assert_eq!(next_truncation_len(200), None);
-    }
-
-    #[test]
-    fn is_context_length_error_matches_known_provider_wordings() {
-        // The exact Ollama wording seen in production (see app log).
-        assert!(is_context_length_error(
-            "Ollama 500 Internal Server Error: {\"error\":\"the input length exceeds the context length\"}"
-        ));
-        assert!(is_context_length_error(
-            "openai: this model's maximum context length is 8192 tokens"
-        ));
-        // OpenAI's plain "too long" wording.
-        assert!(is_context_length_error("openai: Input text is too long"));
-        // Gemini's real over-length wording.
-        assert!(is_context_length_error(
-            "gemini: the input token count (12000) exceeds the maximum number of tokens allowed (8192)."
-        ));
-        assert!(is_context_length_error(
-            "gemini: request too large — try a smaller resume/job ad."
-        ));
-        assert!(!is_context_length_error(
-            "gemini: invalid or unauthorized API key."
-        ));
-        assert!(!is_context_length_error(
-            "Ollama unreachable: connection refused"
-        ));
-        // 404 model-not-found must never be mistaken for a length overflow.
-        assert!(!is_context_length_error(
-            "gemini: model or endpoint not found — models/text-embedding-004 is not found"
-        ));
-    }
-
-    /// `AppHandle`-free fake — succeeds once the (truncated) text is at or
-    /// below `success_at_or_below` chars, otherwise reports a context-length
-    /// overflow. Lets `embed_adaptive`'s retry loop be exercised with no
-    /// network/provider at all. `success_len_sum` accumulates the length of
-    /// every SUCCESSFUL call only — the direct measure of how much of the
-    /// original document actually got embedded, as opposed to `last_len`
-    /// (the most recent attempt, success or failure).
-    struct FakeEmbedAttempt {
-        success_at_or_below: usize,
-        calls: std::sync::atomic::AtomicUsize,
-        last_len: std::sync::atomic::AtomicUsize,
-        success_len_sum: std::sync::atomic::AtomicUsize,
-    }
-
-    impl FakeEmbedAttempt {
-        fn new(success_at_or_below: usize) -> Self {
-            Self {
-                success_at_or_below,
-                calls: std::sync::atomic::AtomicUsize::new(0),
-                last_len: std::sync::atomic::AtomicUsize::new(0),
-                success_len_sum: std::sync::atomic::AtomicUsize::new(0),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl EmbedAttempt for FakeEmbedAttempt {
-        async fn attempt(&self, text: &str) -> AppResult<(Vec<f64>, Usage)> {
-            let len = text.chars().count();
-            self.calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-            self.last_len
-                .store(len, std::sync::atomic::Ordering::SeqCst);
-            if len <= self.success_at_or_below {
-                self.success_len_sum
-                    .fetch_add(len, std::sync::atomic::Ordering::SeqCst);
-                Ok((vec![0.1, 0.2, 0.3], Usage::default()))
-            } else {
-                Err(AppError::Provider(
-                    "mock 500: the input length exceeds the context length".to_string(),
-                ))
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_retries_and_succeeds_on_shorter_input() {
-        let attempt = FakeEmbedAttempt::new(3000);
-        let text = "a".repeat(8000);
-        let result = embed_adaptive(&attempt, &text, 8000, &mut Usage::default()).await;
-        assert!(result.is_ok());
-        // 8000 (fail) -> 4000 (fail) -> 2000 (succeeds, <= 3000). The
-        // successful 2000-char cap then STICKS for the rest of this chunk
-        // too (M1 fix) instead of discarding the remaining 6000 chars: 3
-        // more successful 2000-char sends. 2 failures + 4 successes = 6.
-        assert_eq!(attempt.calls.load(std::sync::atomic::Ordering::SeqCst), 6);
-        assert_eq!(
-            attempt.last_len.load(std::sync::atomic::Ordering::SeqCst),
-            2000
-        );
-        // The WHOLE document was embedded (4 x 2000), not just its first
-        // successful-cap's worth.
-        assert_eq!(
-            attempt
-                .success_len_sum
-                .load(std::sync::atomic::Ordering::SeqCst),
-            8000
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_never_drops_the_unsent_remainder_of_a_chunk() {
-        // The reported production scenario: cap=8000 (Ollama's default), a
-        // document long enough to split into several top-level chunks, and
-        // a provider whose REAL token window only accepts much shorter input
-        // (nomic-embed-text's 2048-token window on a token-dense language).
-        // Before the fix, once a smaller cap succeeded, everything past that
-        // smaller cap within the SAME top-level chunk was silently dropped
-        // forever — a 24,000-char document would embed only 6,000 chars
-        // (3 chunks x first-successful-2000) while still being tagged as a
-        // complete, indexed vector.
-        let attempt = FakeEmbedAttempt::new(2000); // provider accepts <= 2000 chars
-        let text = "a".repeat(24_000);
-        let result = embed_adaptive(&attempt, &text, 8000, &mut Usage::default()).await;
-        assert!(result.is_ok());
-        assert_eq!(
-            attempt
-                .success_len_sum
-                .load(std::sync::atomic::Ordering::SeqCst),
-            24_000,
-            "every char of the document must eventually be embedded"
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_learns_the_working_cap_once_not_per_chunk() {
-        // 3 top-level 8000-char chunks (24,000 / 8000, no growth needed),
-        // provider only accepts <= 2000 chars. WITHOUT hoisting the learned
-        // cap out of `embed_chunk_adaptive`, every chunk independently
-        // re-pays the 8000 -> 4000 -> 2000 discovery ladder (2 wasted failing
-        // calls each = 6 total); WITH it, only chunk 1 pays that cost —
-        // chunks 2 and 3 start straight at the already-learned 2000-char cap.
-        let attempt = FakeEmbedAttempt::new(2000);
-        let text = "a".repeat(24_000);
-        let result = embed_adaptive(&attempt, &text, 8000, &mut Usage::default()).await;
-        assert!(result.is_ok());
-        // chunk 1: 2 failures (8000, 4000) + 4 successes (2000 chars x4) = 6.
-        // chunk 2 and chunk 3: 4 successes each, ZERO failures (learned) = 8.
-        // 6 + 8 = 14 — NOT the 18 it would be if every chunk re-discovered
-        // the cap independently.
-        assert_eq!(attempt.calls.load(std::sync::atomic::Ordering::SeqCst), 14);
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_does_not_let_a_chunks_remainder_poison_the_learned_cap() {
-        // The reported production bug, specifically on the `bounded_split_cap`
-        // GROWTH path (document long enough to need more than
-        // `MAX_CHUNKS_PER_DOCUMENT` chunks at the nominal cap — the growth
-        // ratio doesn't divide evenly by the provider's real limit, so every
-        // chunk ends with a short remainder). A chunk's tiny final remainder
-        // used to permanently shrink the "learned" cap the NEXT chunk started
-        // at, collapsing the whole rest of the document toward 1-char sends
-        // and hitting the attempt ceiling with only a sliver actually
-        // embedded. 300,000 chars at cap=8000 needs 38 chunks (> 32), so
-        // `bounded_split_cap` grows the chunk size to 9,375; the provider
-        // only accepts <= 5,000 chars, forcing exactly one real discovery
-        // (9375 fails, halves to 4687, which succeeds) — that 4687 must
-        // carry forward, not the 1-char tail left over after it.
-        let attempt = FakeEmbedAttempt::new(5000);
-        let text = "a".repeat(300_000);
-        let result = embed_adaptive(&attempt, &text, 8000, &mut Usage::default()).await;
-        assert!(
-            result.is_ok(),
-            "must complete, not hit the {MAX_TOTAL_EMBED_ATTEMPTS}-call ceiling"
-        );
-        assert_eq!(
-            attempt
-                .success_len_sum
-                .load(std::sync::atomic::Ordering::SeqCst),
-            300_000,
-            "the whole document must be embedded"
-        );
-        // Chunk 1: [9375 FAIL, 4687 OK, 4687 OK, 1 OK] = 4 calls, learns 4687.
-        // Chunks 2-32 (31 more): each starts AT the learned 4687, so
-        // [4687 OK, 4687 OK, 1 OK] = 3 calls, ZERO re-discovery failures.
-        // 4 + 31*3 = 97 — nowhere near the un-fixed near-200-and-abort.
-        assert_eq!(
-            attempt.calls.load(std::sync::atomic::Ordering::SeqCst),
-            97,
-            "the learned cap must carry across chunks, not be re-discovered near-per-char"
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_aborts_with_a_clear_error_past_the_total_attempt_ceiling() {
-        // A document long enough that even the DISCOVERED (floor) cap still
-        // needs hundreds of successful sends to fully cover it — this must
-        // abort with a clear error rather than making the caller
-        // (`ai_reembed_all`'s cancellation check runs BETWEEN documents, not
-        // within one) unresponsive for however long that would take. 25
-        // top-level 8,000-char chunks (200,000 / 8,000, no growth needed);
-        // only the floor length (500 chars) ever succeeds, so covering the
-        // whole document would need ~400 successful sends plus the initial
-        // discovery failures — far past the 200-call ceiling.
-        let attempt = FakeEmbedAttempt::new(EMBED_TRUNCATION_FLOOR_CHARS);
-        let text = "a".repeat(200_000);
-        let err = embed_adaptive(&attempt, &text, 8000, &mut Usage::default())
-            .await
-            .unwrap_err();
-        let msg = err.to_string();
-        assert!(
-            msg.contains(&MAX_TOTAL_EMBED_ATTEMPTS.to_string()),
-            "error should cite the ceiling: {msg}"
-        );
-        assert!(msg.to_ascii_lowercase().contains("provider calls"));
-        // Must stop EXACTLY at the ceiling, never run past it.
-        assert_eq!(
-            attempt.calls.load(std::sync::atomic::Ordering::SeqCst),
-            MAX_TOTAL_EMBED_ATTEMPTS
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_bounds_chunk_count_for_a_pathologically_large_document() {
-        // A 2 MB document at cap=8000 would be 250 SEQUENTIAL top-level
-        // chunks (each up to 3 HTTP attempts x up to 5 adaptive-halving
-        // steps worst case) — it was 1 call before this whole feature
-        // existed. `usize::MAX` as the success threshold means every attempt
-        // succeeds immediately, isolating the CHUNK-COUNT bound from the
-        // halving logic tested elsewhere.
-        let attempt = FakeEmbedAttempt::new(usize::MAX);
-        let text = "a".repeat(2_000_000);
-        let result = embed_adaptive(&attempt, &text, 8000, &mut Usage::default()).await;
-        assert!(result.is_ok());
-        let calls = attempt.calls.load(std::sync::atomic::Ordering::SeqCst);
-        assert!(calls <= 32, "expected at most 32 calls, got {calls}");
-        assert_eq!(
-            attempt
-                .success_len_sum
-                .load(std::sync::atomic::Ordering::SeqCst),
-            2_000_000,
-            "the whole document must still be embedded, not truncated to fit the chunk cap"
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_gives_up_at_the_floor_with_a_clear_error() {
-        // A threshold below the floor never succeeds — the loop must stop.
-        let attempt = FakeEmbedAttempt::new(100);
-        let text = "a".repeat(8000);
-        let err = embed_adaptive(&attempt, &text, 8000, &mut Usage::default())
-            .await
-            .unwrap_err();
-        let msg = err.to_string();
-        assert!(
-            msg.contains(&EMBED_TRUNCATION_FLOOR_CHARS.to_string()),
-            "error should mention the floor length: {msg}"
-        );
-        assert!(msg.to_ascii_lowercase().contains("too long"));
-        // 8000 -> 4000 -> 2000 -> 1000 -> 500 (floor): 5 attempts, then give up.
-        assert_eq!(attempt.calls.load(std::sync::atomic::Ordering::SeqCst), 5);
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_never_inflates_a_short_documents_reported_length() {
-        // A 300-char document (well under the 8000-char provider cap) must
-        // never be reported/retried against a cap it was never actually sent
-        // at. Before the fix, `cap` halved 8000→4000→2000→1000→500 while the
-        // ACTUAL bytes sent stayed 300 the whole time (5 byte-identical
-        // requests), and the give-up error falsely said "500 characters".
-        let attempt = FakeEmbedAttempt::new(100); // never succeeds — forces a give-up
-        let text = "a".repeat(300);
-        let err = embed_adaptive(&attempt, &text, 8000, &mut Usage::default())
-            .await
-            .unwrap_err();
-        let msg = err.to_string();
-        assert!(
-            msg.contains("300"),
-            "error must cite the ACTUAL sent length: {msg}"
-        );
-        for phantom in ["8000", "4000", "2000", "1000"] {
-            assert!(
-                !msg.contains(phantom),
-                "error must not cite a cap the document was never sent at: {msg}"
-            );
-        }
-        // 300 chars is already at/below the floor (500) — no retries are
-        // possible, so exactly ONE request is made, not five.
-        assert_eq!(attempt.calls.load(std::sync::atomic::Ordering::SeqCst), 1);
-        assert_eq!(
-            attempt.last_len.load(std::sync::atomic::Ordering::SeqCst),
-            300
-        );
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_does_not_retry_a_non_context_length_error() {
-        struct AlwaysUnauthorized(std::sync::atomic::AtomicUsize);
-
-        #[async_trait]
-        impl EmbedAttempt for AlwaysUnauthorized {
-            async fn attempt(&self, _text: &str) -> AppResult<(Vec<f64>, Usage)> {
-                self.0.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                Err(AppError::Config(
-                    "gemini: invalid or unauthorized API key.".to_string(),
-                ))
-            }
-        }
-
-        let attempt = AlwaysUnauthorized(std::sync::atomic::AtomicUsize::new(0));
-        let err = embed_adaptive(&attempt, "short text", 8000, &mut Usage::default())
-            .await
-            .unwrap_err();
-        assert!(matches!(err, AppError::Config(_)));
-        assert_eq!(attempt.0.load(std::sync::atomic::Ordering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_leaves_partial_usage_in_the_output_param_when_a_later_chunk_fails() {
-        // Simulates a 429 on chunk 3 of 4 (a non-context-length error, never
-        // retried): the FIRST 2 chunks' REAL provider-reported usage must
-        // still be visible to the caller via the `usage` OUTPUT param, never
-        // silently discarded just because the overall call ultimately
-        // failed — `embed_text` records this before propagating the error,
-        // so a partial embed no longer drops already-billed spend from the
-        // ledger (the pre-chunking code made exactly one call per document,
-        // so a failure billed nothing to begin with; this restores that
-        // "never lose real usage" property for the new multi-chunk path).
-        struct FailAfterNSuccesses {
-            succeed_calls: usize,
-            calls: std::sync::atomic::AtomicUsize,
-        }
-
-        #[async_trait]
-        impl EmbedAttempt for FailAfterNSuccesses {
-            async fn attempt(&self, _text: &str) -> AppResult<(Vec<f64>, Usage)> {
-                let n = self.calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                if n < self.succeed_calls {
-                    Ok((
-                        vec![0.1, 0.2],
-                        Usage {
-                            input_tokens: 100,
-                            output_tokens: 0,
-                        },
-                    ))
-                } else {
-                    Err(AppError::RateLimited("mock 429".to_string()))
-                }
-            }
-        }
-
-        let attempt = FailAfterNSuccesses {
-            succeed_calls: 2,
-            calls: std::sync::atomic::AtomicUsize::new(0),
-        };
-        let text = "a".repeat(40); // cap=10 -> 4 top-level chunks, no growth
-        let mut usage = Usage::default();
-        let result = embed_adaptive(&attempt, &text, 10, &mut usage).await;
-        assert!(
-            result.is_err(),
-            "the 3rd chunk's failure must still propagate"
-        );
-        assert_eq!(
-            usage.input_tokens, 200,
-            "usage from the 2 chunks that succeeded before the failure must survive"
-        );
-    }
-
-    // ── split_into_chunks / l2_normalize (pure helpers) ─────────────────────
-
-    #[test]
-    fn split_into_chunks_covers_the_whole_text_without_dropping_a_tail() {
-        let text = "a".repeat(25);
-        let chunks = split_into_chunks(&text, 10);
-        assert_eq!(chunks, vec!["a".repeat(10), "a".repeat(10), "a".repeat(5)]);
-        // Every char of the original text is present across the chunks.
-        assert_eq!(chunks.concat().chars().count(), 25);
-    }
-
-    #[test]
-    fn split_into_chunks_is_a_single_chunk_when_under_the_cap() {
-        assert_eq!(split_into_chunks("short", 8000), vec!["short"]);
-    }
-
-    #[test]
-    fn split_into_chunks_respects_char_boundaries() {
-        let text = "é".repeat(7); // multi-byte char, 2 bytes each
-        let chunks = split_into_chunks(&text, 3);
-        assert_eq!(chunks, vec!["é".repeat(3), "é".repeat(3), "é".to_string()]);
-    }
-
-    #[test]
-    fn split_into_chunks_of_empty_text_is_one_empty_chunk() {
-        // So the caller still makes its usual single provider call rather
-        // than zero (preserves prior single-empty-call behavior).
-        assert_eq!(split_into_chunks("", 8000), vec![""]);
-    }
-
-    #[test]
-    fn bounded_split_cap_is_a_no_op_within_the_chunk_limit() {
-        // 24,000 chars at cap=8000 is 3 chunks — well under the 32 ceiling.
-        assert_eq!(bounded_split_cap(24_000, 8000), 8000);
-    }
-
-    #[test]
-    fn bounded_split_cap_grows_the_cap_to_stay_within_the_limit_without_dropping_text() {
-        // 2 MB at cap=8000 would need 250 chunks — far over the 32 ceiling.
-        let total = 2_000_000;
-        let cap = bounded_split_cap(total, 8000);
-        let needed = total.div_ceil(cap);
-        assert!(needed <= 32, "grown cap {cap} still needs {needed} chunks");
-        // The grown cap must still cover the WHOLE document across at most
-        // 32 chunks — growing the chunk size, never truncating the document.
-        assert!(cap * 32 >= total);
-    }
-
-    #[test]
-    fn l2_normalize_scales_to_unit_length() {
-        let mut v = vec![3.0, 4.0]; // 3-4-5 triangle
-        l2_normalize(&mut v);
-        assert!((v[0] - 0.6).abs() < 1e-9);
-        assert!((v[1] - 0.8).abs() < 1e-9);
-        let norm = (v[0] * v[0] + v[1] * v[1]).sqrt();
-        assert!((norm - 1.0).abs() < 1e-9);
-    }
-
-    #[test]
-    fn l2_normalize_is_a_no_op_on_an_all_zero_vector() {
-        let mut v = vec![0.0, 0.0];
-        l2_normalize(&mut v);
-        assert_eq!(v, vec![0.0, 0.0]);
-    }
-
-    // ── embed_adaptive: chunk-and-mean-pool (the whole-document fix) ────────
-
-    /// Returns a distinct, caller-scripted vector per call (in call order), so
-    /// a test can verify BOTH which text each chunk actually received and how
-    /// the resulting per-chunk vectors were combined.
-    struct SequencedEmbedAttempt {
-        call_texts: std::sync::Mutex<Vec<String>>,
-        vectors: Vec<Vec<f64>>,
-    }
-
-    #[async_trait]
-    impl EmbedAttempt for SequencedEmbedAttempt {
-        async fn attempt(&self, text: &str) -> AppResult<(Vec<f64>, Usage)> {
-            let mut texts = self.call_texts.lock().unwrap();
-            let i = texts.len();
-            texts.push(text.to_string());
-            let v = self
-                .vectors
-                .get(i)
-                .cloned()
-                .expect("test provided fewer scripted vectors than chunks");
-            Ok((
-                v,
-                Usage {
-                    input_tokens: 10,
-                    output_tokens: 0,
-                },
-            ))
-        }
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_embeds_every_chunk_of_a_long_document_not_just_its_prefix() {
-        // cap=10, a 25-char document -> 3 chunks (10 + 10 + 5). A naive single
-        // truncation would have sent ONLY the first 10 chars and silently
-        // dropped the rest while still tagging the result as "complete".
-        let text = format!("{}{}{}", "a".repeat(10), "b".repeat(10), "c".repeat(5));
-        let attempt = SequencedEmbedAttempt {
-            call_texts: std::sync::Mutex::new(Vec::new()),
-            vectors: vec![vec![1.0, 0.0], vec![0.0, 1.0], vec![1.0, 1.0]],
-        };
-
-        let mut usage = Usage::default();
-        let values = embed_adaptive(&attempt, &text, 10, &mut usage)
-            .await
-            .unwrap();
-
-        let texts = attempt.call_texts.lock().unwrap();
-        assert_eq!(texts.len(), 3, "the whole document must be embedded");
-        assert_eq!(*texts, vec!["a".repeat(10), "b".repeat(10), "c".repeat(5)]);
-
-        // Mean of [1,0], [0,1], [1,1] = [2/3, 2/3] -> L2-normalized both
-        // components are equal and the result has unit length.
-        assert!((values[0] - values[1]).abs() < 1e-9);
-        let norm = (values[0] * values[0] + values[1] * values[1]).sqrt();
-        assert!(
-            (norm - 1.0).abs() < 1e-9,
-            "pooled vector must be L2-normalized"
-        );
-
-        // REAL usage summed across every chunk call — 3 chunks x 10 tokens.
-        assert_eq!(usage.input_tokens, 30);
-    }
-
-    #[tokio::test]
-    async fn embed_adaptive_single_chunk_document_is_returned_unpooled() {
-        // Under the cap -> one chunk -> the provider's own vector passes
-        // through mean-pooling as a no-op, then gets L2-normalized.
-        let attempt = SequencedEmbedAttempt {
-            call_texts: std::sync::Mutex::new(Vec::new()),
-            vectors: vec![vec![3.0, 4.0]],
-        };
-        let mut usage = Usage::default();
-        let values = embed_adaptive(&attempt, "short doc", 8000, &mut usage)
-            .await
-            .unwrap();
-        assert_eq!(attempt.call_texts.lock().unwrap().len(), 1);
-        assert!((values[0] - 0.6).abs() < 1e-9);
-        assert!((values[1] - 0.8).abs() < 1e-9);
-        assert_eq!(usage.input_tokens, 10);
     }
 }
